convert pages to typescript
