+++ src/components/TwoPlayerPage.tsx
@@ -1,30 +1,26 @@
 import React from 'react';
 import { useState } from 'react';
 import HexButton from './button/HexButton';
-import SelectNumberButton from './button/SelectNumberButton.js';
-import ColorButton from './button/ColorButton';
+import SelectNumberButton from './button/SelectNumberButton';
 import RestartButton from './button/RestartButton';
 import PlayerTurn from './labels/PlayerTurn';
 import  border from './border/borders';
 import  checkWinBoardPlayer1  from '.././utility/RedGameOverCheck.js';
 import  checkWinBoardPlayer2  from '.././utility/BlueGameOverCheck.js';
-import moveAI from '../utility/AI/AI.js';
 import { COLORS, NOT_ALLOWED_COLOR } from '../constants/colors.js';
 
 export default function TwoPlayerPage() {
-    const create2DArray = (dimension) => {
-        const array2D = [];
+    const create2DArray = (dimension: number) => {
+        const array2D: number[][] = [];
         for (let i = 0; i < dimension; i++) {
             array2D.push(Array(dimension).fill(0)); // Create a new row filled with 0s
         }
         return array2D;
     };
     
-    const [difficulty, selectDifficulty] = useState([true, false, false]);
     const [gameInProgress, setGameInProgress] = useState(false);
     const [twoPlayerBoardDimension, setTwoPlayerBoardDimension] = useState(5);
     const [selectedBoardSize, setSeletedBoardSize] = useState([true, false, false, false])
-    const [selectedColor, setSelectedColor] = useState([true, false]);
     const CENTER_INDEX = Math.floor(twoPlayerBoardDimension/2);
     const [hexagons, setHexagons] = useState(() => {
         const initialHexagons = create2DArray(twoPlayerBoardDimension);
@@ -42,41 +38,10 @@ export default function TwoPlayerPage() {
         setPlayerTurn("Red's Move!")
         const initialHexagons = create2DArray(twoPlayerBoardDimension);
         initialHexagons[CENTER_INDEX][CENTER_INDEX] = -1;
-        if (selectedColor[1]) //selected blue
-        {
-          //move red AI, AIPlayerNumber index is different
-          moveAI({board:initialHexagons, AIPlayerNumber:1, difficulty: difficulty})
-          setRedIsNext(false);
-          if (initialHexagons[CENTER_INDEX][CENTER_INDEX] == -1)
-            initialHexagons[CENTER_INDEX][CENTER_INDEX] = 0;
-        }
-        setHexagons(initialHexagons);          
+        setHexagons(initialHexagons);
     }
 
-    const handleSelectBlueClick = () => {
-      if (gameInProgress)
-      {
-        alert("please restart the game to update color");
-        return;
-      }
-      setSelectedColor([false, true]);
-      moveAI({board:hexagons, AIPlayerNumber:1, difficulty: difficulty})
-      setGameInProgress(true);
-      setRedIsNext(false);
-      if (hexagons[CENTER_INDEX][CENTER_INDEX] == -1)
-        hexagons[CENTER_INDEX][CENTER_INDEX] = 0;
-    }
-
-    const handleSelectRedClick = () => {
-      if (gameInProgress)
-      {
-        alert("please restart the game to update color");
-        return;
-      }
-      setSelectedColor([true, false]);
-    }
-
-    const handleBoardSizeClick = (boardDimension) => {
+    const handleBoardSizeClick = (boardDimension: number) => {
         if (gameInProgress)
         {
             alert("please restart the game to update board size");
@@ -103,31 +68,12 @@ export default function TwoPlayerPage() {
             setSeletedBoardSize([false, false, false, true]);
     }
 
-    const handleDifficultyClick = (difficulty) => {
-        if (gameInProgress)
-        {
-            alert("please restart the game to update difficulty");
-            return;
-        }
+    const handleHexagonClick = (i: number, j: number) => {
+        if (hexagons[i][j] != 0 || gameOver)
+            return
 
-        if (difficulty == 1)
-            selectDifficulty([true, false, false]);
-        else if (difficulty == 2)
-            selectDifficulty([false, true, false]);
-        else 
-            selectDifficulty([false, false, true]);
-    }
-
-    const handleHexagonClick = (i, j) => {
-        if (hexagons[i][j] != 0 
-          || selectedColor[0] == true && !redIsNext
-          || selectedColor[1] == true && redIsNext
-          || gameOver)
-          {
-            return;
-          }
-            
         setGameInProgress(true);
+
         const nextHexagons = JSON.parse(JSON.stringify(hexagons));
 
         if (nextHexagons[CENTER_INDEX][CENTER_INDEX] == -1)
@@ -138,18 +84,10 @@ export default function TwoPlayerPage() {
             if (checkWinBoardPlayer1({hexagons:nextHexagons}))
             {
                 setGameOver(true);
-                setPlayerTurn("Red wins!");
+                setPlayerTurn("Red wins!")
             }      
             else 
-            {             
-              moveAI({board:nextHexagons, AIPlayerNumber:2, difficulty: difficulty})
-              setRedIsNext(true);   
-              if (checkWinBoardPlayer2({board:nextHexagons}))
-              {
-                  setGameOver(true);
-                  setPlayerTurn("Blue wins!");  
-              }    
-            }
+                setPlayerTurn("Blue's Move!")         
         }         
         else 
         {            
@@ -157,19 +95,13 @@ export default function TwoPlayerPage() {
             if (checkWinBoardPlayer2({board:nextHexagons}))
             {
                 setGameOver(true);
-                setPlayerTurn("Blue wins!");
+                setPlayerTurn("Blue wins!")     
             }
             else 
-            {
-              moveAI({board:nextHexagons, AIPlayerNumber:1, difficulty:difficulty})
-              setRedIsNext(false);
-              if (checkWinBoardPlayer1({hexagons:nextHexagons}))
-              {
-                setGameOver(true);
-                setPlayerTurn("Red wins!");
-              }      
-            }            
+                setPlayerTurn("Red's Move!")
         }      
+        
+        setRedIsNext(!redIsNext);
         setHexagons(nextHexagons); 
     };
 
@@ -202,62 +134,28 @@ export default function TwoPlayerPage() {
                <h4> Select Board Size</h4>
                <SelectNumberButton 
                     key={`5-size`}
-                    label={5}
+                    label={'5'}
                     selected={selectedBoardSize[0]}
                     onClick={() => handleBoardSizeClick(5)}
                 />         
                 <SelectNumberButton 
                     key={`7-size`}
-                    label={7}
+                    label={'7'}
                     selected={selectedBoardSize[1]}
                     onClick={() => handleBoardSizeClick(7)}
                 />   
                 <SelectNumberButton 
                     key={`9-size`}
-                    label={9}
+                    label={'9'}
                     selected={selectedBoardSize[2]}
                     onClick={() => handleBoardSizeClick(9)}
                 />      
-                <SelectNumberButton 
+                 <SelectNumberButton 
                     key={`11-size`}
-                    label={11}
+                    label={'11'}
                     selected={selectedBoardSize[3]}
                     onClick={() => handleBoardSizeClick(11)}
-                />         
-               <h4> Select Color</h4>
-               <ColorButton 
-                    key={`redColorButton`}
-                    label={'Red'}
-                    selected={selectedColor[0]}
-                    red = { true }
-                    onClick={() => handleSelectRedClick()}
-                />  
-                <ColorButton 
-                    key={`blueColorButton`}
-                    label={'Blue'}
-                    selected={selectedColor[1]}
-                    red = { false }
-                    onClick={() => handleSelectBlueClick()}
-                />  
-               <h4> Select Difficulty</h4>  
-              <SelectNumberButton 
-                    key={`1-difficulty`}
-                    label={1}
-                    selected={difficulty[0]}
-                    onClick={() => handleDifficultyClick(1)}
-              />      
-              <SelectNumberButton 
-                  key={`2-difficulty`}
-                  label={2}
-                  selected={difficulty[1]}
-                  onClick={() => handleDifficultyClick(2)}
-              />      
-              <SelectNumberButton 
-                  key={`3-difficulty`}
-                  label={3}
-                  selected={difficulty[2]}
-                  onClick={() => handleDifficultyClick(3)}
-              />      
+                />              
                <PlayerTurn 
                     key={`turnLabel`}
                     text={playerTurn}
@@ -265,10 +163,9 @@ export default function TwoPlayerPage() {
                <RestartButton 
                     key={`restartButton`}
                     onClick={() => handleRestartClick()}
-                />
-               
+                />              
             </div>
-            <div display ="inline-block" >
+            <div >
                 <svg viewBox='0 0 1000 800'>
                     {border({borderNumber:0, boardDimension:twoPlayerBoardDimension}) }
                     {border({borderNumber:1, boardDimension:twoPlayerBoardDimension}) }
