--- src/components/labels/PlayerTurn.js
@@ -1,12 +1,0 @@
-import React from 'react';
-
-export default class RestartButton extends React.Component
-{
-    render() {
-        const { text } = this.props;
-
-        return (   
-             <p id = "playerTurn">{text}</p>
-        );
-    }
-}
