--- src/constants/colors.js
@@ -1,3 +1,0 @@
-//unclaimedPiece, player1Piece, player2Piece, player1PieceWon, player2PieceWon
-export const COLORS = ['#A9A9A9', '#E42217', '#7690ac', '#800000', '#000080'];
-export const NOT_ALLOWED_COLOR = '#696969';
