+++ src/constants/colors.ts
@@ -1,0 +1,3 @@
+//unclaimedPiece, player1Piece, player2Piece, player1PieceWon, player2PieceWon
+export const COLORS: string[] = ['#A9A9A9', '#E42217', '#7690ac', '#800000', '#000080'];
+export const NOT_ALLOWED_COLOR: string = '#696969';
